Metrics are collected for process indicators that are not persisted
