+++ central/processindicator/datastore/datastore_impl.go
@@ -102,7 +102,6 @@ func (ds *datastoreImpl) AddProcessIndicators(ctx context.Context, indicators ..
 				return err
 			}
 		} else {
-			recordProcessIndicatorsBatchAdded(identifierBatch)
 			log.Debugf("successfully added a batch of %d process indicators", len(identifierBatch))
 		}
 	}
